Should now extract stats.
After discussion with John and looking over the ARM manuals and gem5
stat output, the script should now be almost ready for running. It still
needs to manage the fact that the stats are not completely in order,
that some stats are occasionally absent, and the code needs to be copied
to the littleCluster section as well.