Data aggregate script works. Code needs tidying.
This script collects all the relevant data from the gem5 stats and
outputs a CSV. I have tested that it works on the roi-out-2. It seems
the file size is reduced by about a factor of 1000, probably due to the
simulator recording way more information than we can get using PMUs.